requirements.txt
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
imageio==2.25.1
kiwisolver==1.4.4
matplotlib==3.7.0
networkx==3.0
numpy==1.24.2
opencv-python==4.7.0.68
packaging==23.0
patchify==0.2.3
Pillow==9.4.0
pydicom==2.3.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
scikit-image==0.19.3
scipy==1.10.1
six==1.16.0
tifffile==2023.2.3
